requirements.txt
langchain
langchain-huggingface
langchain_community
langchain-google-community
streamlit
langchainhub
google-api-python-client
python-dotenv
huggingface_hub
transformers